+++ src/CEC.hpp
@@ -21,30 +21,30 @@ namespace gmum {
     boost::shared_ptr<std::vector<unsigned int> > assignment;
     boost::shared_ptr<arma::mat> points;
     boost::shared_ptr<Algorithm> algorithm;
-    const float killThreshold;
+    const double killThreshold;
   public:
     CEC(boost::shared_ptr<arma::mat> points, 
 	boost::shared_ptr<std::vector<unsigned int> > assignment, 
 	boost::shared_ptr<Algorithm> algorithm,
-	float killThreshold, std::vector<ClusterType> type,
-	std::vector<float> radius, std::vector<arma::mat> covMatrices);
+	double killThreshold, std::vector<ClusterType> type,
+	std::vector<double> radius, std::vector<arma::mat> covMatrices);
 
     CEC(boost::shared_ptr<arma::mat> points, 
 	boost::shared_ptr<std::vector<unsigned int> > assignment, 
 	boost::shared_ptr<Algorithm> algorithm,
-	float killThreshold, int numberOfClusters);
+	double killThreshold, int numberOfClusters);
 
     // for debug purposes
     std::vector<boost::shared_ptr<Cluster> > clusters;
     void loop();
     void singleLoop();
-    float entropy();
+    double entropy();
     std::vector<unsigned int> getAssignment();
     std::vector<arma::rowvec> centers();
     std::vector<arma::mat> cov();
     unsigned int iters();
     std::list<unsigned int> getNrOfClusters();
-    std::list<float> getEnergy();
+    std::list<double> getEnergy();
   };
 
 }
